rest-api/requirements.txt
fastapi>=0.115,<0.116
uvicorn>=0.34,<0.35
sqlalchemy>=2.0,<2.1
asyncpg>=0.30,<0.31
passlib[argon2]>=1.7,<1.8
alembic>=1.15,<1.16
python-jose[cryptography]>=3.4,<3.5
pydantic>=2.11,<2.12
pydantic-settings>=2.9,<2.10
black>=25.1.0,<25.2
pyright>=1.1,<1.2
ruff>=0.11,<0.12
isort>=6.0,<6.1
aiofiles>=24.1,<24.2